Fixed bug in module code

+++ code.py
@@ -117,8 +117,7 @@ class CodeImageIterator(object):
     
     def allPixels(self):
         i,j = numpy.mgrid[0:self.imageSize[1],0:self.imageSize[0]]
-        return (i % self.stepsize == 0) & (j % self.stepsize == 0)
-        #return numpy.fromfunction(lambda i,j: ((i % self.stepsize == 0) * (j % self.stepsize == 0)), (self.imageSize[1],self.imageSize[0]))
+        return ((i % self.stepsize == 0) & (j % self.stepsize == 0)).astype(float)
 
     def lookupPixel(self, word):
         """
